LayoutTests/imported/w3c/web-platform-tests/css/css-tables/tentative/tbody-height-redistribution-expected.txt

TBODY/THEAD/TFOOT height redistribution.

FF passes most of the tests.
Legacy Chrome does not distribute TBODY height to rows.
At the time this test was written, there was no spec. Currently, spec is in the process of being written.
1) Empty row group's size is 0 with collapsed borders

2) Empty row group's size is 0 with separate borders.

3) Row group height is redistributed to unconstrained rows in proportion to their size.

0,0
0,0
4) Constrained fixed rows do not grow if there are unconstrained rows

0,1
5) Constrained percentage are resolved against row group height.

0,0
0,0
0,2
6) Row group height is treated as min-height.

0,0
7) Constrained/unconstrained mix: row group height is distributed to unconstrained cells.

0,0
0,1
0,2
0,3
8) Unconstrained empty rows grow to fill table if all other rows are constrained. Not in FF.

0,0
0,1

0,1
0,2
9) Empty section height is not ignored

10) Section with empty rows is not considered empty.

In FF, empty section height is 100, but table's height is 0?

11) Mixed constrained and unconstrained empty rows: only unconstrained rows grow.


FAIL table 1 assert_equals:
<table id="one" style="border-collapse: collapse;border-spacing: 10px 10px;border:10px solid green" data-expected-width="0" data-expected-height="0">
  <tbody>
    <tr> </tr>
  </tbody>
</table>
width expected 0 but got 10
FAIL table 2 assert_equals:
<table id="two" style="border-collapse: separate;border-spacing: 10px 10px;border:10px solid green" data-expected-width="20" data-expected-height="20">
  <tbody>
    <tr> </tr>
  </tbody>
</table>
width expected 20 but got 40
FAIL table 3 assert_equals:
<table id="three">
  <tbody style="height:100px" data-expected-height="100">
    <tr data-expected-height="25">
      <td><div style="height:10px">0,0</div></td>
    </tr>
    <tr data-expected-height="75">
      <td><div style="height:30px">0,0</div></td>
    </tr>
  </tbody>
</table>
height expected 100 but got 40
FAIL table 4 assert_equals:
<table id="four">
  <tbody style="height:100px">
    <tr style="height: 10px" data-expected-height="10">
      <td></td>
    </tr>
    <tr data-expected-height="90">
      <td>0,1</td>
    </tr>
  </tbody>
</table>
height expected 90 but got 18
FAIL table 5 assert_equals:
<table id="five">
  <tbody style="height:100px">
    <tr style="height:25%" data-expected-height="25">
      <td>0,0</td>
    </tr>
    <tr style="height:50%" data-expected-height="50">
      <td>0,0</td>
    </tr>
    <tr data-expected-height="25">
      <td>0,2</td>
    </tr>
  </tbody>
</table>
height expected 25 but got 18
PASS table 6
FAIL table 7 assert_equals:
<table id="seven">
  <tbody style="height:100px">
    <tr style="height:20px" data-expected-height="20"><td>0,0</td></tr>
    <tr style="height:30px" data-expected-height="30"><td>0,1</td></tr>
    <tr data-expected-height="25"><td>0,2</td></tr>
    <tr data-expected-height="25"><td>0,3</td></tr>
  </tbody>
</table>
height expected 25 but got 18
FAIL table 8 assert_equals:
<table id="eight">
  <tbody style="height:100px">
    <tr style="height:20px" data-expected-height="20"><td>0,0</td></tr>
    <tr style="height:30px" data-expected-height="30"><td>0,1</td></tr>
    <tr data-expected-height="50"><td></td></tr>
  </tbody>
</table>
height expected 50 but got 0
FAIL table 9 assert_equals:
<table id="eightplus">
  <tbody style="height:100px">
    <tr data-expected-height="50"><td></td></tr>
    <tr style="height:20px" data-expected-height="20"><td>0,1</td></tr>
    <tr style="height:30px" data-expected-height="30"><td>0,2</td></tr>
  </tbody>
</table>
height expected 50 but got 0
FAIL table 10 assert_equals:
<table id="nine">
  <tbody style="height:75px" data-expected-height="75">
  </tbody>
</table>
height expected 75 but got 0
FAIL table 11 assert_equals:
<table id="nine">
  <tbody style="height:75px" data-expected-height="75">
    <tr data-expected-height="25"></tr>
    <tr data-expected-height="25"></tr>
    <tr data-expected-height="25"></tr>
  </tbody>
</table>
height expected 75 but got 0
FAIL table 12 assert_equals:
<table id="ten" style="width:50px">
  <tbody style="height:50px" data-expected-height="50">
    <tr style="height:0" data-expected-height="0"><td></td></tr>
    <tr data-expected-height="25"><td></td></tr>
    <tr data-expected-height="25"><td></td></tr>
  </tbody>
</table>
height expected 50 but got 0
